+++ apps/common/ospray_testing/builders/ClippingGeometries.cpp
@@ -63,7 +63,8 @@ cpp::Group ClippingGeometries::buildGroup() const
   spheresGeometry.commit();
 
   cpp::GeometricModel model(spheresGeometry);
-  if (rendererType == "pathtracer" || rendererType == "scivis") {
+  if (rendererType == "pathtracer" || rendererType == "scivis"
+      || rendererType == "ao") {
     cpp::Material material(rendererType, "obj");
     material.setParam("kd", vec3f(.1f, .4f, .8f));
     material.commit();