+++ apps/common/ospray_testing/builders/ParticleVolume.cpp
@@ -175,7 +175,8 @@ cpp::Group ParticleVolume::buildGroup() const
 
     cpp::GeometricModel isoModel(isoGeom);
 
-    if (rendererType == "pathtracer" || rendererType == "scivis") {
+    if (rendererType == "pathtracer" || rendererType == "scivis"
+        || rendererType == "ao") {
       cpp::Material mat(rendererType, "obj");
       mat.setParam("kd", vec3f(1.f));
       mat.setParam("d", 0.5f);