+++ apps/common/ospray_testing/builders/PerlinNoiseVolumes.cpp
@@ -106,7 +106,8 @@ cpp::GeometricModel createGeometricModel(
 {
   cpp::GeometricModel geometricModel(geo);
 
-  if (rendererType == "pathtracer" || rendererType == "scivis") {
+  if (rendererType == "pathtracer" || rendererType == "scivis"
+      || rendererType == "ao") {
     cpp::Material objMaterial(rendererType, "obj");
     objMaterial.setParam("kd", kd);
     objMaterial.commit();