+++ apps/ospExamples/GLFWOSPRayWindow.h
@@ -19,6 +19,7 @@ enum class OSPRayRendererType
 {
   SCIVIS,
   PATHTRACER,
+  AO,
   DEBUGGER,
   OTHER
 };