+++ apps/ospTestSuite/test_geometry.cpp
@@ -29,7 +29,7 @@ void SpherePrecision::SetUp()
 
   renderer.setParam("pixelSamples", 16);
   renderer.setParam("backgroundColor", vec4f(0.2f, 0.2f, 0.4f, 1.0f));
-  if (rendererType == "scivis") {
+  if (rendererType == "scivis" || rendererType == "ao") {
     renderer.setParam("aoSamples", 16);
     renderer.setParam("aoIntensity", 1.f);
   } else if (rendererType == "pathtracer") {
@@ -115,7 +115,7 @@ INSTANTIATE_TEST_SUITE_P(TestScenesGeometry,
                            "subdivision_cube",
                            "cornell_box_photometric",
                            "planes"),
-        ::testing::Values("scivis", "pathtracer")));
+        ::testing::Values("scivis", "pathtracer", "ao")));
 
 INSTANTIATE_TEST_SUITE_P(TestScenesClipping,
     FromOsprayTesting,
@@ -128,7 +128,7 @@ INSTANTIATE_TEST_SUITE_P(TestScenesClipping,
                            "clip_with_bspline_curves",
                            "clip_gravity_spheres_volume",
                            "clip_perlin_noise_volumes"),
-        ::testing::Values("scivis", "pathtracer")));
+        ::testing::Values("scivis", "pathtracer", "ao")));
 
 TEST_P(FromOsprayTestingDirect, test_scenes)
 {